Make username@host in prompt more useful
Since both my laptop and my development VM are called 'jrose.*',
showing the first segment of the hostname after my username is
pretty silly looking, and doesn't tell me which I'm on.

This extracts the first bit of the hostname that doesn't match
my username, and exports it as $UNIQUEHOST, so I can pull it
into my prompt and see something like "jrose@laptop" or "jrose@vm"
instead.